requirements.txt
# mysql
mysql-connector-python==8.0.31

